+++ kedro_vertexai/config.py
@@ -1,7 +1,8 @@
-from typing import Optional, Dict
-from pydantic import BaseModel
+import os
+from typing import Dict, List, Optional
 
-from kedro.config import MissingConfigException
+from pydantic import BaseModel
+from pydantic.networks import IPvAnyAddress
 
 DEFAULT_CONFIG_TEMPLATE = """
 # Configuration used to run the pipeline
@@ -31,12 +32,24 @@ run_config:
   # volume after pipeline finishes) [in seconds]. Default: 1 week
   ttl: 604800
 
+  # Optional network configuration
+  # network:
+
+    # Name of the vpc to use for running Vertex Pipeline
+    # vpc: my-vpc
+
+    # Hosts aliases to be placed in /etc/hosts when pipeline is executed
+    # host_aliases:
+    #  - ip: 127.0.0.1
+    #    hostnames: me.local
+
   # What Kedro pipeline should be run as the last step regardless of the
   # pipeline status. Used to send notifications or raise the alerts
   # on_exit_pipeline: notify_via_slack
 
-  # Optional section allowing adjustment of the resources
-  # reservations and limits for the nodes
+  # Optional section allowing adjustment of the resources, reservations and limits
+  # for the nodes. When not provided they're set to 500m cpu and 1024Mi memory.
+  # If you don't want to specify pipeline resources set both to None in __default__.
   resources:
 
     # For nodes that require more RAM you can increase the "memory"
@@ -49,10 +62,6 @@ run_config:
       cpu: 8
       memory: 1Gi
 
-    # GPU-capable nodes can request 1 GPU slot
-    tensorflow_step:
-      nvidia.com/gpu: 1
-
     # Default settings for the nodes
     __default__:
       cpu: 200m
@@ -60,9 +69,19 @@ run_config:
 """
 
 
+class HostAliasConfig(BaseModel):
+    ip: IPvAnyAddress
+    hostnames: List[str]
+
+
 class ResourcesConfig(BaseModel):
     cpu: Optional[str]
     memory: Optional[str]
+
+
+class NetworkConfig(BaseModel):
+    vpc: Optional[str]
+    host_aliases: Optional[List[HostAliasConfig]] = []
 
 
 class RunConfig(BaseModel):
@@ -72,7 +91,7 @@ class RunConfig(BaseModel):
     description: Optional[str]
     experiment_name: str
     scheduled_run_name: Optional[str]
-    service_account: Optional[str]
+    network: Optional[NetworkConfig] = NetworkConfig()
     ttl: int = 3600 * 24 * 7
     resources: Optional[Dict[str, ResourcesConfig]] = dict(
         __default__=ResourcesConfig(cpu="500m", memory="1024Mi")
@@ -81,94 +100,33 @@ class RunConfig(BaseModel):
     def resources_for(self, node):
         if node in self.resources.keys():
             result = self.resources["__default__"].dict()
-            result.update({k: v for k, v in self.resources[node].dict().items() if v is not None})
+            result.update(
+                {
+                    k: v
+                    for k, v in self.resources[node].dict().items()
+                    if v is not None
+                }
+            )
             return result
         else:
             return self.resources["__default__"].dict()
-
-
-class NetworkConfig(BaseModel):
-    vpc: str
 
 
 class PluginConfig(BaseModel):
     project_id: str
     region: str
     run_config: RunConfig
+    service_account: Optional[str]
 
+    @staticmethod
+    def sample_config(**kwargs):
+        return DEFAULT_CONFIG_TEMPLATE.format(**kwargs)
 
-
-# class Config(object):
-#     def __init__(self, raw):
-#         self._raw = raw
-#
-#     def _get_or_default(self, prop, default):
-#         return self._raw.get(prop, default)
-#
-#     def _get_or_fail(self, prop):
-#         if prop in self._raw.keys():
-#             return self._raw[prop]
-#         else:
-#             raise MissingConfigException(
-#                 f"Missing required configuration: '{self._get_prefix()}{prop}'."
-#             )
-#
-#     def _get_prefix(self):
-#         return ""
-#
-#     def __eq__(self, other):
-#         if isinstance(other, Config):
-#             return self._raw == other._raw
-#         else:
-#             return False
-
-    @property
-    def on_exit_pipeline(self):
-        return self._get_or_default("on_exit_pipeline", None)
-
-    @property
-    def vertex_ai_networking(self):
-        return VertexAiNetworkingConfig(
-            self._get_or_default("vertex_ai_networking", {})
-        )
-
-    @property
-    def node_merge_strategy(self):
-        strategy = str(self._get_or_default("node_merge_strategy", "none"))
-        if strategy not in ["none", "full"]:
-            raise ValueError(
-                f"Invalid {self._get_prefix()}node_merge_strategy: {strategy}"
-            )
-        else:
-            return strategy
-
-    def _get_prefix(self):
-        return "run_config."
-
-
-# class PluginConfig(Config):
-#     @property
-#     def run_config(self) -> RunConfig:
-#         cfg = self._get_or_fail("run_config")
-#         return RunConfig(cfg)
-#
-#     @staticmethod
-#     def sample_config(**kwargs):
-#         return DEFAULT_CONFIG_TEMPLATE.format(**kwargs)
-#
-#     @property
-#     def project_id(self):
-#         return self._get_or_fail("project_id")
-#
-#     @property
-#     def region(self):
-#         return self._get_or_fail("region")
-#
-#     @staticmethod
-#     def initialize_github_actions(project_name, where, templates_dir):
-#         os.makedirs(where / ".github/workflows", exist_ok=True)
-#         for template in ["on-push.yml"]:
-#             file_path = where / ".github/workflows" / template
-#             template_file = templates_dir / f"github-{template}"
-#             with open(template_file, "r") as tfile, open(file_path, "w") as f:
-#                 f.write(tfile.read().format(project_name=project_name))
+    @staticmethod
+    def initialize_github_actions(project_name, where, templates_dir):
+        os.makedirs(where / ".github/workflows", exist_ok=True)
+        for template in ["on-push.yml"]:
+            file_path = where / ".github/workflows" / template
+            template_file = templates_dir / f"github-{template}"
+            with open(template_file, "r") as tfile, open(file_path, "w") as f:
+                f.write(tfile.read().format(project_name=project_name))